Refactor podcasts router into modular package
