+++ backend/api/routers/podcasts/__init__.py
@@ -1,0 +1,13 @@
+from fastapi import APIRouter
+
+router = APIRouter(prefix="/podcasts", tags=["Podcasts (Shows)"])
+
+from .crud import router as crud_router
+from .distribution import router as distribution_router
+from .spreaker import router as spreaker_router
+
+router.include_router(crud_router)
+router.include_router(distribution_router)
+router.include_router(spreaker_router)
+
+__all__ = ["router"]
